修复部分问题

+++ app/src/main/java/com/bai/chesscard/bean/BeanCharge.java
@@ -8,5 +8,5 @@ package com.bai.chesscard.bean;/**
  */
 public class BeanCharge {
     public int type;
-    public int amount;
+    public long amount;
 }
